auth_service/requirements.txt
sqlalchemy 
psycopg2-binary 
alembic 
grpcio 
grpcio-tools
pytest
passlib
PyJWT
bcrypt==4.0.1